refactor(mm-next): move getSectionFromPremiumHeaderData
