+++ packages/mirror-media-next/pages/category/[slug].js
@@ -7,9 +7,11 @@ import {
   fetchHeaderDataInDefaultPageLayout,
   fetchHeaderDataInPremiumPageLayout,
   getPostsAndPostscountFromGqlData,
-  getSectionFromPremiumHeaderData,
 } from '../../utils/api'
-import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
+import {
+  getSectionAndTopicFromDefaultHeaderData,
+  getSectionFromPremiumHeaderData,
+} from '../../utils/data-process'
 import { setPageCache } from '../../utils/cache-setting'
 import Layout from '../../components/shared/layout'
 import { Z_INDEX } from '../../constants/index'
